pyproject.toml
[tool.poetry]
name = "loquendobot"
version = "0.1.0"
description = ""
authors = ["ekisu <dts.ramon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.0.1"
gtts = "^2.5.1"
pydub = "^0.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
